Looking to split out logic for processing for zapxml
